Split up parallel tree nodes tests.
