+++ tests/test_parallel_tree_nodes_1.py
@@ -25,28 +25,3 @@ test_script = os.path.join(script_path, "parallel/parallel_tree_nodes.py")
 class ParallelTreeNodesTest1(TempDirTest, ParallelTest):
     test_script = test_script
     args = ("all", "forest", 0, 0)
-
-
-class ParallelTreeNodesTest2(TempDirTest, ParallelTest):
-    test_script = test_script
-    args = ("all", "tree", 0, 0)
-
-
-class ParallelTreeNodesTest3(TempDirTest, ParallelTest):
-    test_script = test_script
-    args = ("all", "prog", 0, 0)
-
-
-class ParallelTreeNodesTest4(TempDirTest, ParallelTest):
-    test_script = test_script
-    args = ("all", "forest", 2, 0)
-
-
-class ParallelTreeNodesTest5(TempDirTest, ParallelTest):
-    test_script = test_script
-    args = ("all", "forest", 0, 1)
-
-
-class ParallelTreeNodesTest6(TempDirTest, ParallelTest):
-    test_script = test_script
-    args = ("all", "nodes-10", 0, 0)
